refactor: 일관성 향상을 위한 수정
