+++ backend/main/src/main/java/kr/huichan/bak/main/document/Navbar.java
@@ -1,0 +1,12 @@
+package kr.huichan.bak.main.document;
+
+import org.springframework.data.annotation.Id;
+
+import kr.huichan.bak.main.dto.NavbarMenu;
+
+public record Navbar(
+    @Id
+    String id,
+    NavbarMenu menu
+    ) {
+}
