+++ backend/main/src/main/java/kr/huichan/bak/main/document/User.java
@@ -1,4 +1,4 @@
-package kr.huichan.bak.main.dto;
+package kr.huichan.bak.main.document;
 
 import java.util.Collections;
 import java.util.HashSet;
@@ -12,7 +12,7 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 @Document(collection = "users")
-public class UserDocument implements UserDetails {
+public class User implements UserDetails {
     @Id
     private ObjectId _id;
     private String username;
@@ -32,7 +32,7 @@ public class UserDocument implements UserDetails {
         return this.authorities;
     }
 
-    public UserDocument(String username, String password) {
+    public User(String username, String password) {
         this.username = username;
         this.password = password;
         this.authorities = new HashSet<>(Collections.singletonList(new SimpleGrantedAuthority("USER")));
