--- backend/main/src/main/java/kr/huichan/bak/main/dto/NavbarMenuDocument.java
@@ -1,10 +1,0 @@
-package kr.huichan.bak.main.dto;
-
-import org.springframework.data.annotation.Id;
-
-public record NavbarMenuDocument(
-    @Id
-    String id,
-    NavbarMenu menu
-    ) {
-}
