+++ backend/main/src/main/java/kr/huichan/bak/main/repository/NavbarRepository.java
@@ -4,9 +4,9 @@ import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import kr.huichan.bak.main.dto.NavbarMenuDocument;
+import kr.huichan.bak.main.document.Navbar;
 
 @Repository
-public interface NavbarRepository extends MongoRepository<NavbarMenuDocument, ObjectId> {
+public interface NavbarRepository extends MongoRepository<Navbar, ObjectId> {
     
 }
